requirements.txt
openai==1.30.1
langchain==0.2.1
langchain-openai==0.1.6
streamlit==1.36.0y

yt-dlp==2024.6.21
openai-whisper==20231117
ffmpeg-python==0.2.0
python-dotenv==1.0.1